More documentation! Got rid of warnings of unchecked calls to CommunicationRequest.
Thinking about looking at whether SendMessageThread is actually needed, since requests are always sent intentionally. Initial worry was that there may be issues with the JavaFX thread. If not, could instead have each Client have a data field for ObjectOutputStream, in which case Client may have to become a class as opposed to an interface.